Transformations

AlgorithmAnalysis.convex_interpolation — Function

convex_interpolation(opt::Node)

Given an optimization node, replaces all convex functions with their interpolation conditions.

AlgorithmAnalysis.smooth_convex_interpolation — Function

smoot_convex_interpolation(opt::Node)

Given an optimization node, replaces all smooth convex functions with their interpolation conditions.

AlgorithmAnalysis.sector_bounded_interpolation — Function

sector_bounded_interpolation(opt::Node)

Given an optimization node, replaces all sector bounded functions with their interpolation conditions.

AlgorithmAnalysis.gram_transformation — Function

gram_transformation(opt::Node)

Given an optimization node, for each vector space, replaces all vectors in the space with the condition that their Gram matrix is positive semidefinite. All inner products are flattened into new symbolic variables.

AlgorithmAnalysis.lyapunov_transformation — Function

lyapunov_transformation(node)

Given a Lyapunov certificate node, constructs an optimization problem that searches for a valid Lyapunov certificate of convergence. The Lyapunov candidate is linear in the algorithm state, where the state is specified by transitions: $V(x) = \theta ⋅ x$. The analysis then uses the S-procedure to search for the parameters $\theta$ such that the Lyapunov candidate satisfies the following two conditions: - Positivity: $V(x) \geq \text{performance}$ - Decreasing: $V(x₊) \leq \text{rate}\,V(x)$ where the performance measure and rate are specified by the node.